Fix inks again in README.md.
I didn't fix them right last time. The `kwiver/` part is not necessary within a readme file. The new links are tested to work when the readme is displayed on github.

Paths in the command line examples were changed similarly but the reason was stylistic: `</path/to/kwiver/source>` should be contrued to include `kwiver/`.